Update patch set 1
Patch Set 1:

Yeah, and it is.
You either start it and never touch it again (current setup), or never start it at all. If it starts and stops, the RPM firmware will shut down the device a short time afterwards, even if you did restart the daemon. The only 8974 I've seen so far in which this apparently doesn't happen is the hammerhead.

Patch-set: 1