Update patch set 1
Patch Set 1:

I was going to submit this a week or so ago, but the problem is that the overlay here was either not used or it was subsequently overwritten by the overlay in /vendor/cm.

Also, shouldn't we increase the values for maxMessageSize, maxMessageWidth and maxMessageHeight for the G2?

Patch-set: 1
Reviewer: m4570d0n <4381@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0