Update patch set 2
Patch Set 2: -Code-Review

this works fine as is, but based on what Marko Zidari? told me in an email, this can also be applied to voice-speaker as well, and does indeed improve in-call audio quality.  i'll push up an updated PS tomorrow to include that for testing as well.

Patch-set: 2