Update patch set 1
Patch Set 1:

so whats the suggestion here? remove the configs for now, re-add them later if needed? and adjust the init script to not call isrc_util with sec_config? its already a bit of a mess in the commited code... and generating an error each time we boot is not the right strategy...?

Patch-set: 1