Update patch set 1
Patch Set 1:

I agree with Georg, the config for the G2 is an absolute mess (no offense to however wrote it). And since it's pretty much the same hardware as the N5, it probably should be structurally inline with it.

Patch-set: 1
Reviewer: Rashed Abdel-Tawab <3962@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0