Update patch set 1
Patch Set 1:

Should be noted this patch is useless without: http://review.cyanogenmod.org/#/c/84385/

I didn't add it to BoardConfig because g2 is not defining any of their .te's in there. Which still puzzles me, apparently its supposed to implicitly include them in the union, but I didn't have success with that approach on at least one of my devices.

Patch-set: 1