Update patch set 1
Patch Set 1:

UNION is a global variable. If a file is included elsewhere (system or qcom, in this case), adding it to a device's UNION is a no-op (it'll end up de-duped out). All files in the UNION variable are searched in all of the defined DIRs. The only way this won't work is if you're accidentally resetting the UNION (:= vs +=) variable along the way, or are including in the wrong order.

If a filename is new, you do need to add it though.

Patch-set: 1