Update patch set 2
Patch Set 2:

Yes I did, but I made a mistake. I checked GPS on maps, and at the time I didn't know it takes the location by A-GPS too. I tested it with GPS tools, and wasn't working. The good thing here, is that the reboots stopped. So, it is 100% gps related. This new patch set, is already working?

Patch-set: 2