Update patch set 3
Patch Set 3:

OK, if you guys have a series of patches that are ready and work, tag them with for-rashed or something. There're like 4 different patches for fixing audio and like 3 for DT2W and I'm a tad bit too busy for the next few days to test everything. Plus I'm hammering out a GPS issue so the time I do have will probably be spent on that. Thanks.

Patch-set: 3