Update patch set 1
Patch Set 1: Code-Review-2

There's a reason I've held off on doing this. We need to add some sort of assert that won't let the user even flash the build if they're on the JB bootloader. While you and I will know that we're using the wrong bootloader if it hangs at the LG logo, the average user won't.

Patch-set: 1