Update patch set 2
Patch Set 2:

You're assuming the user will be able to remotely know what happened. Most users panic when they see a bootloop or a hang. They will have absolutely no way to know what happened, which is why we have to add an assert that tells the user what's wrong. I'll write the code for it tonight or tomorrow.

Patch-set: 2