Update patch set 1
Patch Set 1:

Users on xda are complaing that it is very low. (And I can say that it is...)
We're using on cm the lg stock mixer_paths file, but this file has the levels set for lg framework which has a built-in amp. software.

So, with this values, we get closer to stock levels, without amplification.

Patch-set: 1