Update patch set 5
Patch Set 5:

I'm on a D800 and the audio is fine. It's fine for many people. I asked a question on the CM13 thread (I'm dr4stic, btw) about what variants the people with audio issues had, what carrier, and what region they were in.

I'm also thinking of dabbling with changing some of the voice processing settings for the D800 since it is the only variant with an extra mic, and an external audience chip for noise cancellation. 

One way or another, audio doesn't seem to be standard across the variants.

Patch-set: 5