Update patch set 1
Patch Set 1: Code-Review+1

I provided a flashable build for 12.1 which many swore by. It was not very far removed from the version of Snap as it is now. I posted (and have been using) a build with the patches up through 2016-01-08 and it's been working flawlessly, and many on the forum have downloaded and are using it. Unless the latest spat of patches do anything crazy, I'd say this is ready for prime time. Please approve.

Patch-set: 1
Label: Code-Review=+1