Update patch set 1
Patch Set 1:

Not sure what the original code intended to do, but without this patch, changing the network type does not have any effect. It doesn't work at all. And with this patch, it does. 

And the G2/G3 aren't the only devices implementing an override for this. There are a number of devices which have implemented this on MM. I started playing with code for this after lifting it from a motorola msm8960 device. Somewhere along the line the original code isn't covering all eventualities.

Patch-set: 1