Update patch set 1
Patch Set 1:

Hello Ethan, I do not really understand why we are having this issue. It seems G3 is plagued with the same problem. At the moment I'm trying to fully understand the batch of patches google brought in Android 6 in which the function getRadioCapability got introduced. Do you by coincidence if more devices than G2 and G3 are affected?

Patch-set: 1