Update patch set 1
Patch Set 1:

> Found out that there are a ton of devices which actually use this
 > "Override hack". A quick google with the commit title shows up the
 > devices. Change-id doesn't point to the other devices than only G3.

Yes, there are a number of devices implementing this in cm-13.0, including the LG G3, the HTC M8, devices from Sony and Motorola. This is a fairly well accepted workaround for a real problem. The only difference seems to be that the G2 is being punished for it.

Patch-set: 1