Update patch set 1
Patch Set 1:

Hello Ethan. In this new audio_policy.conf I set the sample_rates from fixed values which we use now to dynamic which is set in this new audio_policy. Setting the fixed sample_rates breaks the compressie_offload for soms reason and tenders the bug where you aren't able to scroll through a song. I'll look at your comment as wel tomorrow. It's just a first setup for now. It will be adjusted according to feedback but for now experience with most people is quite good

Patch-set: 1