Update patch set 1
Patch Set 1:

Looks like it'll be worthwhile to decompile the LG RIL class and see what they have as message ID 130 (RIL_REQUEST_GET_RADIO_CAPABILITY). The LGE implementation is returning something totally weird. We should remap it correctly, or remap it so the RIL will not recognize it and return the not supported error code.

Patch-set: 1