Update patch set 3
Patch Set 3:

Sorry for the bump Ethan, but just wanted to bring up Bacon recently turned it back on after bumping the timer rate. Our values are identical to theirs so it should run better than when it was last enabled. Only relevant diff is probably cpufreq, so if we merge some interactive and cpufreq updates it may be better. It does give a nice performance boost though, which is why I want to find the source of instability and fix it in the near future.

Patch-set: 3
Reviewer: Tyler Dunn <12592@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0