Update patch set 1
Patch Set 1: Code-Review-1

Obviously a work in progress. I'm inviting everyone that I feel could contribute/offer an opinion on what route to take, or even contribute and edit it. -1 until we have functional code.

The idea is to get the back LED light by the power button working, and hopefully working on notifications/calls/etc (stock behavior). The back LED is defined as an attention light. At the moment, all this commit does is enable the use of attention lights and create a separate function to handle the attention light. This function along with set_light_attention will need to be modified in order to actually set the brightness/blink/blink time. 

The sysfs paths added are the correct paths and have been named with an attempt to follow the current naming scheme. What we should do in the new function is set blink and blink ms (default is fine though, we can just enable blink) and enable/disable as the other functions do. I'm not sure what the correct step would be here - but this is to play with and see where we can get.

This also opens the door to a setting that would allow users to turn it off per say if they are using a case that blocks it, rendering it useless.

I don't believe all variants have the back LED light though, can someone confirm what variants do/don't?

Patch-set: 1
Label: Code-Review=-1