Update patch set 7
Patch Set 7:

(1 comment)

Also, fixed the static const char[] typing. Sultan did this in http://review.cyanogenmod.org/#/c/141673/2 and I believe it leads to better compiler optimization and typing. I can revert the format changes if you like as well. The commit at this point is practically the same as Sultan's but at first I didn't want to nix the attention light code on the fact that we may one day work with it, but there have been no complaints and unfortunately most G2 cases cover the power button (the one I own does) :( If we did work on the attention lights code it would likely have to be changed/rewritten so we might as well nix it. I'm open to all suggestions though, of course.

Patch-set: 7