Update patch set 1
Patch Set 1: Code-Review+1

For clarification: The original code forces the camera preview image to an aspect ratio of 16:9 (1920x1280). Most of the camera's native resolutions are not in 16:9 ratio, and therefore the preview is actually a cropped form of the final picture. The end user cannot see the full frame of the photo before it's taken.


Forcing the preview resolution results in an always on "cooler looking" appearance, what with transparencies and all, however it makes the camera less functional as a camera. This change reverts (removes) the forced 1920x1280 preview resolution.

Patch-set: 1