Update patch set 2
Patch Set 2:

Hello Ethan, Atm. I uploaded this commit for safekeeping and others to experiment with. I'll put up a testbuild with this first to try and gather some usefull feedback and also try to exactly figure out the up/downsides of this commit. I'll be researching this fluence behaviour in the meantime. I'm aware I removed the false parameter, why is it in there if it's disabled in the first place (Same with audiorec)?

Patch-set: 2