Update patch set 1
Patch Set 1: Code-Review-1

Not sure what this does, but I get 1610 mA using the regular unmodified nightlies and either of a QC 2.0 capable charger or my Samsung tablet charger. I doubt we have the capability for QC 2.0 because I think there's an additional hardware component necessary, as well as some licensing. We (and the G3 for that matter) are not on this list: https://www.qualcomm.com/documents/quick-charge-device-list

Patch-set: 1
Label: Code-Review=-1