Update patch set 2
Patch Set 2:

https://github.com/CyanogenMod/android_device_lge_g2-common/blob/cm-13.0/rootdir/etc/init.g2.rc#L149

We do have it. The point is it wasn't 'missing' as per the commit title. It was written that way with the intention of being defined as yes or not being defined at all. It's a really harmless warning that literally does nothing but create one line in logs. No other devices to my knowledge define it as it's pretty pointless to define.

Patch-set: 2