Update patch set 3
Patch Set 3:

I'd be more than willing to push as this is a nice improvement, however can you link to the original commit as well? iirc we synced up at one point I thought...

Also, I posted the git line you needed to use. If that was not your latest commit, then git checkout and the SHA of the commit, then push again and it will update the patch set.

Patch-set: 3