Update patch set 3
Patch Set 3:

> Tested on D805, signal strength has dropped a bit though. At least
 > is a light in the tunnel, thanks for this decision. I was actually
 > a bit scared about using a different RIL class. I'll test it on
 > marshmallow too.

Which version did you test on if not Marshmallow? Also, is the reception actually lower (are you dropping calls or seeing less bandwidth) or are you just going by the bars in the notification? If I had to guess I can't imagine the ACTUAL reception is different, but rather the reporting. The physical reception should be a function of the radio firmware. And for me, the ril has never quite worked properly on MM. Dialing 611 is a nightmare, and holding/switching calls can occasionally go completely wrong.

Patch-set: 3