Update patch set 3
Patch Set 3: Code-Review-1

Still doesn't address everything in the discussion from the cm-13.0 commit. audio_policy.msm8974 is redundant, for example. Also, this is growing in terms of how many changes are made. It would be good to split some of these changes up per the issue they address.

Patch-set: 3