Update patch set 1
Patch Set 1:

> It's up to you what you feel the better solution is. I mean, we
 > could either go back to the old size, or use this/disable some
 > kernel things. I just felt it was better to adjust the sizes to
 > their actual sizes (other than userdata and system).

I dont want to "go back to the old size" if it's incorrect. If this has to be moved down to variant trees, that's fine too. This should be correct, not whatever works for the build.

There are other solutions too (such as the one posed here) along with https://review.cyanogenmod.org/#/c/174166/ or https://review.cyanogenmod.org/#/c/164667/

Patch-set: 1