Update patch set 1
Patch Set 1:

> > That would prevent a factory reset that was initiated from
 > android
 > > though wouldn't it?
 > 
 > As TWRP is currently built, for us at least, /misc is not in fstab
 > or mounted, so anything in /misc is ignored.  Presumably the ROM
 > sends the commands to the boot command-line, since everything works
 > for us, as-is, without /misc

Right, but without hacking on common repos or shipping our own bootloader message lib to point at a file on /cache/ (like I've done on tenderloin since it has no extra partition to use for this), this is the best option. Since we have available, unused, partitions, it make no sense not to use them.

Patch-set: 1