Update patch set 1
Patch Set 1:

> > Has this been tested at all?  I am inclined to adopt it for next
 > > TWRP build.  Alternatively, I am inclined to use the BOOTLOADER
 > > OFFSET to offset past the wipe command issued by the recovery key
 > > combo.
 > 
 > ooh, i didn't think of leveraging the offset. that's a good idea.
 > yea, it works, i've been using it for awhile

 > > Has this been tested at all?  I am inclined to adopt it for next
 > > TWRP build.  Alternatively, I am inclined to use the BOOTLOADER
 > > OFFSET to offset past the wipe command issued by the recovery key
 > > combo.
 > 
 > ooh, i didn't think of leveraging the offset. that's a good idea.
 > yea, it works, i've been using it for awhile

Matt Mower didn't like the idea, but in my estimation that might be the better move.  Then we're on the "correct" partition and we skip over the unwanted command.  The caveat obviously being the ROM needs to use the same offset, but that same caveat applies to ROM with this patch - ROM would need to mount pad as /misc.

Patch-set: 1