Update patch set 1
Patch Set 1:

> We can do that? Why haven't we done it yet? That's a much better
 > fix rather than using /pad

The offset was designed for htc devices who's bootloader actually writes to misc with an offset (on htc you can boot to recovery, but also boot to recovery and wipe, so we're ok listening to bootloader commands on them). I never thought of using it for this purpose but it should work fine.

Patch-set: 1