Update patch set 2
Patch Set 2:

Hey, it's been a long time since I popped in here but I just got around to booting up my G2 for the first time in about half a year if not longer (shifted to Pixel and stock).

That aside, I do like the initial change. Was there a difference in logging with just the xtra3 files? Was the locking tested across all devices? Same question on https servers, do you have a list of countries it doesn't work in? I would assume XTRA3 would be the default on newer phones so it would be weird if it didn't work across the board. If it's a USA thing, we could always shift back to individualized gps config files (although I hate the sound of that).

I'll get around to catching up on G2 and Lineage news shortly, and then test out some GPS changes too. Do you have any numbers available already? It'd be nice to see some benchmarked numbers in commit messages in the future (eg, lock time before: 0.257ms, lock time after: 0.113ms, Tested On: d800 d801 d802).

Patch-set: 2